Safari-611.1.10/LayoutTests/editing/pasteboard/copy-paste-bidi-expected.txt

These tests to see if bidi text with a base writing direction of right to left is put onto the paste board in logical (DOM) order.

first test - before:
| "1234שדגכ:"

first test - after:
| "1234שדגכ:<#selection-caret>"

second test - before:
| "\n"
| <div>
|   "1234שדגכ:"
| "\n"
| <div>
|   style="direction: rtl;"
|   "1234שדגכ:"
| "\n"

second test - after:
| "\n"
| <div>
|   <div>
|     "1234שדגכ:"
|   <div>
|     style="direction: rtl;"
|     "1234שדגכ:<#selection-caret>"
| "\n"
